Post_Training_Quantization/PyTorch/pyproject.toml
[project]
name = "pytorch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dahuffman>=0.4.2",
    "fvcore>=0.1.5.post20221221",
    "numpy>=2.3.1",
    "onnx-graphsurgeon>=0.5.8",
    "onnxsim>=0.4.36",
    "torch>=2.7.1",
    "ultralytics>=8.3.162",
]
